removed unused WeatherDataStore class
