--- Weather/Service/Router/unused/WeatherDataStore.swift
@@ -1,27 +1,0 @@
-//
-//  WeatherDataStore.swift
-//  Weather
-//
-//  Created by Nischal Hada on 6/22/18.
-//  Copyright © 2018 NischalHada. All rights reserved.
-//
-
-import Foundation
-
-protocol WeatherData {
-    func getBulkWeatherInformation(data : [String:Any]) -> [WeatherInformation]?
-}
-
-class WeatherDataStore:WeatherData {
-    func getBulkWeatherInformation(data : [String:Any]) -> [WeatherInformation]?{
-        guard let list = data["list"] as? [AnyObject]else{
-            return nil
-        }
-        var arrayWeather: [WeatherInformation] = []
-        for properties in list{
-            let result = WeatherInformation.init(json: properties as? [String : Any])
-            arrayWeather.append(result!)
-        }
-        return arrayWeather
-    }
-}
